requirements.txt
pyrtlsdr==0.3.0
numpy==2.4.0
scipy==1.16.3
setuptools
flask==3.1.2
flask-socketio==5.6.0
eventlet==0.40.4